changed datatypes

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,15 +11,15 @@
 int main(void)
 {
 	long first = 0;
-	int second = 1;
-	int next;
+	long second = 1;
+	long next;
 	int i = 0;
 	int input = 50;
 
 	for (i = 2 ; i < input ; i++)
 	{
 		next = first + second;
-		printf("%d%s", next, (i < input) ? ", " : "");
+		printf("%ld%s", next, (i < input) ? ", " : "");
 		first = second;
 		second = next;
 	}
